feat(self): merge dashboard and administration
Fixes #1429
